5458: divide historical balance block into new component
