+++ src/ducks/Signals/signalFormManager/signalCrudForm/formParts/TriggerMetricTypesResolver.js
@@ -1,26 +1,47 @@
 import React from 'react'
 import TriggerFormAssetWallet from './TriggerFormAssetWallet'
 import TriggerFormTrendingWordsTypes from './TriggerFormTrendingWordsTypes'
-import { TRENDING_WORDS } from '../../../utils/constants'
+import { ETH_WALLET, TRENDING_WORDS } from '../../../utils/constants'
+import TriggerFormHistoricalBalance from './TriggerFormHistoricalBalance'
 
 const TriggerMetricTypesResolver = ({
-  isNew,
   address,
   values,
   metaFormSettings,
   setFieldValue
 }) => {
-  const { metric } = values
+  const { metric, target } = values
 
-  const isTrendingWords = metric.value === TRENDING_WORDS
+  let TypeComponent
 
-  const TypeComponent = isTrendingWords
-    ? TriggerFormTrendingWordsTypes
-    : TriggerFormAssetWallet
+  let checkPossibleTarget = () => {
+    if (Array.isArray(target)) {
+      setFieldValue(
+        'target',
+        target.length > 0 ? target[0] : metaFormSettings.target.value
+      )
+    }
+  }
+
+  switch (metric.value) {
+    case TRENDING_WORDS: {
+      TypeComponent = TriggerFormTrendingWordsTypes
+      break
+    }
+    case ETH_WALLET: {
+      checkPossibleTarget()
+      TypeComponent = TriggerFormHistoricalBalance
+      break
+    }
+    default: {
+      checkPossibleTarget()
+      TypeComponent = TriggerFormAssetWallet
+      break
+    }
+  }
 
   return (
     <TypeComponent
-      isNew={isNew}
       values={values}
       byAddress={address}
       metaFormSettings={metaFormSettings}
